+++ packages/api-reference/src/index.ts
@@ -7,6 +7,8 @@ export { default as GettingStarted } from '@/components/GettingStarted.vue'
 export { createApiReference } from '@/standalone/lib/html-api'
 
 export { Sidebar } from '@/components/Sidebar'
+export { useSidebar } from '@/features/sidebar'
+
 export { Card, CardHeader, CardContent, CardFooter, CardTabHeader, CardTab } from '@/components/Card'
 
 // TODO: Ideally, we’d remove those exports or at least not export them through the root index.