+++ integrations/fastify/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @scalar/fastify-api-reference
 
+## 1.31.12
+
+## 1.31.11
+
 ## 1.31.10
 
 ## 1.31.9